+++ src/main/java/lagatrix/server/exceptions/connection/BadClassFormatException.java
@@ -3,13 +3,13 @@ package lagatrix.server.exceptions.connection;
 /**
  * This error is used if the client sends an invalid object.
  *
- * @author javier
- * @since 1.0
+ * @author javierfh03
+ * @since 0.2
  */
 public class BadClassFormatException extends ConnectionException {
 
-    public BadClassFormatException(String description) {
-        super(description);
+    public BadClassFormatException() {
+        super("Unexpected class");
     }
 
 }