+++ src/main/java/lagatrix/server/exceptions/connection/ConnectionException.java
@@ -5,25 +5,12 @@ import lagatrix.server.exceptions.LagatrixException;
 /**
  * This exception represents the errors relates to the connection.
  *
- * @author javier
- * @since 1.0
+ * @author javierfh03
+ * @since 0.2
  */
 public class ConnectionException extends LagatrixException {
 
     public ConnectionException(String description) {
         super(description);
-    }
-
-    /**
-     * This method generate the exception message.
-     * 
-     * @param failClass The class who fail.
-     * @param message Description of action.
-     * @param ip The ip of teh client.
-     * @return The message.
-     */
-    public static String getMessage(Class failClass, String message, String ip) {
-        return String.format("Problem connection because %s in '%s' class with client %s", 
-                message, failClass.getSimpleName(), ip);
     }
 }